move tests like the functions were moved
